Manage rows-columns & some refac

+++ KambanSolution/Kamban/Repository/ColumnInfo.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public byte Order { get; set; }
+        public int Order { get; set; }
         public BoardInfo Board { get; set; }
         public int Width { get; set; }
         public int Height { get; set; }
